# TMF673_GeographicAddress/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference : TMF 673 - Place - Geographic Address Management\n\
    ### Release : 19.5 - January 2020\n\nThe Geographic Address API provides a standardized\
    \ client interface to an Address management system.\nIt allows looking for worldwide\
    \ addresses.\nIt can also be used to validate geographic address data, to be sure\
    \ that it corresponds to a real address.\n\n### Resources\n- GeographicAddress\n\
    - GeographicSubAddress\n- GeographicAddressValidation\n- Hub\n\n### Operations\n\
    Geographic Address Management API performs the following operations :\n- Retrieve\
    \ a geographic address or a collection of geographic addresses\n- Retrieve a geographic\
    \ sub-address\n- Create a geographic address validation\n- Retrieve, update and\
    \ delete an existing  geographic address validation\n- Notify events on these\
    \ resources"
  version: "4.0.1"
  title: "API Place - GeographicAddress"
host: "serverRoot:8080"
basePath: "/tmf-api/geographicAddressManagement/v4/"
...
schemes:
- "https"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
  /geographicAddressValidation:
    get:
      tags:
      - "geographicAddressValidation"
      summary: "List or find GeographicAddressValidation objects"
      description: "This operation list or find GeographicAddressValidation entities"
      operationId: "listGeographicAddressValidation"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GeographicAddressValidation"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "GeographicAddressValidation"
    post:
      tags:
      - "geographicAddressValidation"
      summary: "Creates a GeographicAddressValidation"
      description: "This operation creates a GeographicAddressValidation entity."
      operationId: "createGeographicAddressValidation"
      parameters:
      - in: "body"
        name: "geographicAddressValidation"
        description: "The GeographicAddressValidation to be created"
        required: true
        schema:
          $ref: "#/definitions/GeographicAddressValidation_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/GeographicAddressValidation"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "GeographicAddressValidation"
  /geographicAddressValidation/{id}:
    get:
      tags:
      - "geographicAddressValidation"
      summary: "Retrieves a GeographicAddressValidation by ID"
      description: "This operation retrieves a GeographicAddressValidation entity.\
        \ Attribute selection is enabled for all first level attributes."
      operationId: "retrieveGeographicAddressValidation"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the GeographicAddressValidation"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/GeographicAddressValidation"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "GeographicAddressValidation"
    patch:
      tags:
      - "geographicAddressValidation"
      summary: "Updates partially a GeographicAddressValidation"
      description: "This operation updates partially a GeographicAddressValidation\
        \ entity."
      operationId: "patchGeographicAddressValidation"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the GeographicAddressValidation"
        required: true
        type: "string"
      - in: "body"
        name: "geographicAddressValidation"
        description: "The GeographicAddressValidation to be updated"
        required: true
        schema:
          $ref: "#/definitions/GeographicAddressValidation_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/GeographicAddressValidation"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "GeographicAddressValidation"
...
definitions:
...
  GeographicAddressValidation:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the Address Validation"
      href:
        type: "string"
        description: "An URI used to access to the address validation resource"
      provideAlternative:
        type: "boolean"
        description: "Indicator provided by the requester to specify if alternate\
          \ addresses must be provided in case of partial or fail result."
      validationDate:
        type: "string"
        format: "date-time"
        description: "Date when the address validation is performed"
      validationResult:
        type: "string"
        description: "Result of the address validation (success, partial, fails)"
      alternateGeographicAddress:
        type: "array"
        items:
          $ref: "#/definitions/GeographicAddress"
      state:
        $ref: "#/definitions/TaskStateType"
      submittedGeographicAddress:
        description: "the address as submitted to validation"
        $ref: "#/definitions/GeographicAddress"
      validGeographicAddress:
        description: "the correct form of the validated address in case of validation\
          \ success"
        $ref: "#/definitions/GeographicAddress"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "This resource is used to manage address validation request and response"
    example:
      validationResult: "validationResult"
      '@baseType': "@baseType"
      alternateGeographicAddress:
      - country: "country"
        streetType: "streetType"
        geographicLocation:
          '@referredType': "@referredType"
          '@baseType': "ResourceSpecification"
          '@type': "GeoJsonPoint"
          bbox:
          - 0.80082819046101150206595775671303272247314453125
          - 0.80082819046101150206595775671303272247314453125
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "https://mycsp.com:8080/tmf-api/schema/Resource/LogicalResourceSpecification.schema.json"
        city: "city"
        streetNr: "streetNr"
        '@type': "@type"
        locality: "locality"
        postcode: "postcode"
        streetNrLast: "streetNrLast"
        streetNrSuffix: "streetNrSuffix"
        streetName: "streetName"
        stateOrProvince: "stateOrProvince"
        '@baseType': "@baseType"
        name: "name"
        streetNrLastSuffix: "streetNrLastSuffix"
        streetSuffix: "streetSuffix"
        geographicSubAddress:
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - country: "country"
        streetType: "streetType"
        geographicLocation:
          '@referredType': "@referredType"
          '@baseType': "ResourceSpecification"
          '@type': "GeoJsonPoint"
          bbox:
          - 0.80082819046101150206595775671303272247314453125
          - 0.80082819046101150206595775671303272247314453125
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "https://mycsp.com:8080/tmf-api/schema/Resource/LogicalResourceSpecification.schema.json"
        city: "city"
        streetNr: "streetNr"
        '@type': "@type"
        locality: "locality"
        postcode: "postcode"
        streetNrLast: "streetNrLast"
        streetNrSuffix: "streetNrSuffix"
        streetName: "streetName"
        stateOrProvince: "stateOrProvince"
        '@baseType': "@baseType"
        name: "name"
        streetNrLastSuffix: "streetNrLastSuffix"
        streetSuffix: "streetSuffix"
        geographicSubAddress:
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@type': "@type"
      validGeographicAddress:
        country: "country"
        streetType: "streetType"
        geographicLocation:
          '@referredType': "@referredType"
          '@baseType': "ResourceSpecification"
          '@type': "GeoJsonPoint"
          bbox:
          - 0.80082819046101150206595775671303272247314453125
          - 0.80082819046101150206595775671303272247314453125
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "https://mycsp.com:8080/tmf-api/schema/Resource/LogicalResourceSpecification.schema.json"
        city: "city"
        streetNr: "streetNr"
        '@type': "@type"
        locality: "locality"
        postcode: "postcode"
        streetNrLast: "streetNrLast"
        streetNrSuffix: "streetNrSuffix"
        streetName: "streetName"
        stateOrProvince: "stateOrProvince"
        '@baseType': "@baseType"
        name: "name"
        streetNrLastSuffix: "streetNrLastSuffix"
        streetSuffix: "streetSuffix"
        geographicSubAddress:
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      provideAlternative: true
      submittedGeographicAddress:
        country: "country"
        streetType: "streetType"
        geographicLocation:
          '@referredType': "@referredType"
          '@baseType': "ResourceSpecification"
          '@type': "GeoJsonPoint"
          bbox:
          - 0.80082819046101150206595775671303272247314453125
          - 0.80082819046101150206595775671303272247314453125
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "https://mycsp.com:8080/tmf-api/schema/Resource/LogicalResourceSpecification.schema.json"
        city: "city"
        streetNr: "streetNr"
        '@type': "@type"
        locality: "locality"
        postcode: "postcode"
        streetNrLast: "streetNrLast"
        streetNrSuffix: "streetNrSuffix"
        streetName: "streetName"
        stateOrProvince: "stateOrProvince"
        '@baseType': "@baseType"
        name: "name"
        streetNrLastSuffix: "streetNrLastSuffix"
        streetSuffix: "streetSuffix"
        geographicSubAddress:
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      id: "id"
      href: "href"
      validationDate: "2000-01-23T04:56:07.000+00:00"
      state: {}
      '@schemaLocation': "http://example.com/aeiou"
  GeographicAddressValidation_Create:
    type: "object"
    required:
    - "provideAlternative"
    - "submittedGeographicAddress"
    properties:
      provideAlternative:
        type: "boolean"
        description: "Indicator provided by the requester to specify if alternate\
          \ addresses must be provided in case of partial or fail result."
      submittedGeographicAddress:
        description: "the address as submitted to validation"
        $ref: "#/definitions/GeographicAddress"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "This resource is used to manage address validation request and response\n\
      Skipped properties: id,href,alternateGeographicAddress,validGeographicAddress,state,validationResult,validationDate"
    example:
      '@baseType': "@baseType"
      '@type': "@type"
      provideAlternative: true
      submittedGeographicAddress:
        country: "country"
        streetType: "streetType"
        geographicLocation:
          '@referredType': "@referredType"
          '@baseType': "ResourceSpecification"
          '@type': "GeoJsonPoint"
          bbox:
          - 0.80082819046101150206595775671303272247314453125
          - 0.80082819046101150206595775671303272247314453125
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "https://mycsp.com:8080/tmf-api/schema/Resource/LogicalResourceSpecification.schema.json"
        city: "city"
        streetNr: "streetNr"
        '@type': "@type"
        locality: "locality"
        postcode: "postcode"
        streetNrLast: "streetNrLast"
        streetNrSuffix: "streetNrSuffix"
        streetName: "streetName"
        stateOrProvince: "stateOrProvince"
        '@baseType': "@baseType"
        name: "name"
        streetNrLastSuffix: "streetNrLastSuffix"
        streetSuffix: "streetSuffix"
        geographicSubAddress:
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        - subUnitNumber: "subUnitNumber"
          '@type': "@type"
          buildingName: "buildingName"
          '@baseType': "@baseType"
          levelType: "levelType"
          subUnitType: "subUnitType"
          levelNumber: "levelNumber"
          name: "name"
          subAddressType: "subAddressType"
          id: "id"
          href: "href"
          privateStreetNumber: "privateStreetNumber"
          '@schemaLocation': "http://example.com/aeiou"
          privateStreetName: "privateStreetName"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
...
  TaskStateType:
    type: "string"
    description: "Possible values for the state of a task"
    enum:
    - "accepted"
    - "terminatedWithError"
    - "inProgress"
    - "done"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "string"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "string"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"